Moved loadNext() at AbstractTransactionDownloader level.
